+++ assets/react/components/grid-item.jsx
@@ -8,7 +8,8 @@ function GridItem({variant, description, image, exposed, tileIcons, tileBorders,
         <Wrapper className={exposed ? "exposed" : ""}>
             <Item className={variant} style={{
                 '--background-image': `url(${image}`,
-                '--border-width': tileBorders ? 'var(--tile-border-width)' : 0
+                '--border-width': tileBorders ? 'var(--tile-border-width)' : 0,
+                '--overlay-opacity': tileBorders ? '0.1' : '0'
             }}>
                 {exposed && <ItemDescription style={{
                       '--text-size': `var(--font-size-${exposeFontSize})`
@@ -73,7 +74,7 @@ const Item = styled.div`
     left: 0;
     width: 100%;
     height: 100%;
-    opacity: 0.1;
+    opacity: var(--overlay-opacity);
   }
 
   &.one {